Separate datashader_utilities file
